% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countParsimonyInformative.r
\name{countParsimonyInformative}
\alias{countParsimonyInformative}
\title{Count Parsimony informative sites}
\usage{
countParsimonyInformative(DNAbinObj, calc = "absolute")
}
\arguments{
\item{DNAbinObj}{An alignment in the form of \link{ape}'s DNAbin}

\item{calc}{A string indicating if the calculation should return the absolute count of parsimony informative sites ('absolute') or the percent of parsimony informative sites relative to the length of the alignment ('relative')}
}
\value{
If calc is 'absolute', returns an integer count of parsimony informative sites in the alignment. If calc is 'relative', returns a numeric 0 <= n <= 1 representing the percent of parsimony informative sites relative to the length of the alignment
}
\description{
Counts the parsimony informative sites in a single alignment. The input must be an alignment in the
form of \link{ape}'s DNAbin object. This is primarily an internal function used by \link{getPisCounts}.
Note that ambiguity codes are ignored! The only bases considered are the standard DNA bases: A, G, C, T
}
\seealso{
\link{getPisCounts} \link{getAlignmentsByPIS}
}
