% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentsByPIS.R
\name{getAlignmentsByPIS}
\alias{getAlignmentsByPIS}
\title{Get top alignments by count of parsimony informative sites}
\usage{
getAlignmentsByPIS(pisData, cutoff, outputDir, returnPisData = FALSE)
}
\arguments{
\item{pisData}{A PISdata matrix produced by \link{getPisCounts}}

\item{cutoff}{A numeric 0 < n < 1 or an integer n >= 1. If cutoff is < 1, it is interpreted as requesting the top
loci above the specified percentile as ranked by PIS (e.g. cutoff = 0.8 requests loci with counts of parsimony
informative sites greater than or equal to the 80th percentile). If cutoff is an integer >= 1, it is interpreted as
requesting the top <cutoff> loci (e.g. cutoff = 20 requests the top 20 loci). NOTE: requesting an integer
number of loci returns EXACTLY that many loci even if there are ties (e.g. if you request 20 loci, you get 20 loci
even if loci 21-30 are tied with the 20th ranked locus by counts of parsimony informative sites).}

\item{outputDir}{The directory to which the requested subset of alignments will be copied}

\item{returnPisData}{Boolean. If TRUE, the function returns a PISdata matrix containing only the subset
of loci remaining after filtering. If FALSE, the function returns nothing.}
}
\value{
If returnPisData is TRUE, returns a matrix of class 'PISdata' for only the subset of loci remaining after filtering.
The first column contains paths to alignment files. The second column
contains DNAbin representations of the alignments. The third column contains counts of parsimony informative sites.
The fourth column contains alignment lengths. If returnPisData is false, the function returns nothing and only copies
the subset of alignments remaining after filtering to outputDir.
}
\description{
Uses the PISdata matrix output by \link{getPisCounts} to create a directory containing copies of the top alignments
as ranked by their counts of parsimony informative sites.
}
\seealso{
\link{countParsimonyInformative} \link{getPisCounts}
}
