% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAstral.r
\name{getAstral}
\alias{getAstral}
\title{Download Astral}
\usage{
getAstral(install.dir = getwd(), method)
}
\arguments{
\item{install.dir}{Path to the directory where Astral will be downloaded to. By default, this is the current working directory.}

\item{method}{Method is an optional argument passed to \code{\link[utils:download.file]{utils::download.file()}}; see \code{\link[utils:download.file]{utils::download.file()}} for more info.}
}
\value{
The function returns nothing (void) and is called entirely to download Astral and set the astralPath global option.
}
\description{
Retrieves the Astral v5.7.1 .jar file from github and sets a global option
called 'astralPath' to a be the location of that file. Note that astralPath
can also be passed to the callAstral function manually if you'd prefer to use your
own version of Astral.
}
\seealso{
\link{callAstral}
}
