% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologyDist.R
\name{getNamesForDistances}
\alias{getNamesForDistances}
\title{Get Names For Distances function}
\usage{
getNamesForDistances(combo, treeSet)
}
\arguments{
\item{combo}{a vector containing exactly 2 items: the indices of a pair of trees. Used to select the trees from treeSet and extract their names}

\item{treeSet}{an ape multiPhylo object containing the trees to which the indices in combo refer}
}
\value{
description Returns a distance between the pair of trees
}
\description{
Internal function getting names for the vector returned by toplogyDist for use with parApply
}
\keyword{internal}
