% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPisCounts.r
\name{getPisCounts}
\alias{getPisCounts}
\title{Get counts of parsimony informative sites}
\usage{
getPisCounts(
  alignDir,
  ext,
  calc = "absolute",
  threads = detectCores(),
  format = "fasta"
)
}
\arguments{
\item{alignDir}{Path to the directory containing alignment files}

\item{ext}{The file extension of the alignment files (e.g. "file.nexus" is matched by ext = "nexus"). If not set, all files in the directory are used.}

\item{calc}{A string indicating if the calculation should return the absolute count of parsimony informative sites ('absolute') or the percent of parsimony informative sites relative to the length of the alignment ('relative')}

\item{threads}{Number of simultaneous processes used for counting parsimony infomative sites. By
default, this uses the number of cores returned by \link[parallel:detectCores]{parallel::detectCores}.}

\item{format}{Either 'nexus' or 'fasta' indicating what format the alignments are stored in. The default is 'fasta'}
}
\value{
A matrix of class 'PISdata'. The first column contains paths to alignment files. The second column
contains DNAbin representations of the alignments. The third column contains counts of parsimony informative sites.
The fourth column contains alignment lengths.
}
\description{
Counts the parsimony informative sites in multiple alignments (input as a path to a directory
containing alignments). Returns a matrix of class 'PISdata' that's used by \link{getAlignmentsByPIS}
to subset alignments.
}
\seealso{
\link{countParsimonyInformative} \link{getAlignmentsByPIS}
}
