% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTreesDir.R
\name{readTreesDir}
\alias{readTreesDir}
\title{Read a directory of genetrees in newick format}
\usage{
readTreesDir(treesdir, ext, threads = detectCores())
}
\arguments{
\item{treesdir}{A path the the directory containing the genetrees where each file contains exactly one genetree in newick format}

\item{ext}{The file extension of the alignment files (e.g. "file.treefile" is matched by ext = "treefile"). If not set, all files in the directory are used.}

\item{threads}{Number of simultaneous processes used for reading trees. By
default, this uses the number of cores returned by \link[parallel:detectCores]{parallel::detectCores}.}
}
\value{
An \link{ape} multiPhylo object containing the trees that were read from the directory where each tree is named as the file it was read from
}
\description{
Reads a directory of genetrees in newick format where each file contains one genetree and returns a \link{ape} multiphylo object where each
genetree is named the same as the file it was read from (e.g. when reading a file 'UCEtree01.treefile', the corresponding tree
in the multiphylo object will be named 'UCEtree01.treefile). This function mainly exists to make using \if{html}{\out{<futureFunction>}} easier.
}
\details{
The way \link{ape}'s read.tree() is implemented, text preceding the first '(' is treated as the name of the tree. However, the way
I've written this, I assume there is no text preceding the first '('. If there is, this will result in trees being named like '\if{html}{\out{<filename>}} \if{html}{\out{<precedingText>}}'.
}
\seealso{
\link[ape:read.tree]{ape::read.tree}
}
