% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologyDist.R
\name{topologyDist}
\alias{topologyDist}
\title{Get distances between trees using \link[TreeDist:TreeDistance]{TreeDist::TreeDistance}}
\usage{
topologyDist(treeSet, threads = detectCores(), targetTree, samples)
}
\arguments{
\item{treeSet}{A set of trees as an ape multiPhylo object. See, for example, the \link[ape:read.tree]{ape::read.tree} documentation.}

\item{threads}{Number of simultaneous processes used for calculating tree distances. By
default, this uses the number of cores returned by \link[parallel:detectCores]{parallel::detectCores}.}

\item{targetTree}{Target tree (as an ape 'phylo' object) to which the treeSet is compared. If set, this
changes the behaviour of the function substantially. Instead of returning distances between pairs of trees
in treeSet the function will
return distances between the trees in treeSet and the targetTree.}

\item{samples}{Integer. If not set (default), distance between all possible sets of trees are returned. If set,
a random sample of those distances (with replacement) of size 'samples' is returned instead.}
}
\value{
a vector containing generalized Robinson-Foulds distances
}
\description{
Returns a vector containing generalized Robinson-Foulds distances between trees within a set of trees or between a single
'target' tree and a set of trees using the \link[TreeDist:TreeDistance]{TreeDist::TreeDistance} function.This metric is slightly different
from how Robinson-Foulds distances were originally defined, but it works with trees that don't necessarily
contain exactly the same taxa. Please see the description for the \link[TreeDist:TreeDistance]{TreeDist::TreeDistance} function for a more
thorough explanation. This matrix returned by this function can be used to visualize the effects of filtering trees
by parsimony informative sites
}
