% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologyDist.R
\name{TreeDistWrapper}
\alias{TreeDistWrapper}
\title{TreeDistance Wrapper function}
\usage{
TreeDistWrapper(combo, treeSet)
}
\arguments{
\item{combo}{a vector containing exactly 2 items: the indices of a pair of trees. The distance between this pair is
evaluated by \link[TreeDist:TreeDistance]{TreeDist::TreeDistance}}

\item{treeSet}{an ape multiPhylo object containing the trees to which the indices in combo refer}
}
\value{
description Returns a distance between the pair of trees
}
\description{
Internal \link[TreeDist:TreeDistance]{TreeDist::TreeDistance} Wrapper function for use with parApply; might be avoidable? Matrix of trees rather than indices?
}
\keyword{internal}
