% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCETools_simTree.R
\docType{data}
\name{UCETools_simTree}
\alias{UCETools_simTree}
\title{UCETools simulation tree}
\format{
An object of class \code{phylo} of length 4.
}
\source{
\url{https://github.com/PopGen33/UCETools/}
}
\usage{
UCETools_simTree
}
\description{
\subsection{\code{UCETools_simTree}}{

An \link{ape} phylo object containing the phylogeneitc tree on which the 800 example UCE alignments
included in this package were simulated. The alignments can be accessed via the file path returned by:
\code{system.file("extdata", "UCE_aligns", package = "UCETools")}. The trees associated with these
alignments are can be accessed via the file path returned by: \code{system.file("extdata", "UCE_aligns_genetrees", package = "UCETools")}.
The genetrees were infered using a simple maximum likelihood search using IQTREE.
}
}
\keyword{datasets}
